requirements.txt
setuptools>=68.0.0
Django==5.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-allauth==0.63.3
django-cors-headers==4.3.1
python-decouple==3.8
requests==2.31.0
PyJWT==2.8.0
cryptography==41.0.7
user-agents==2.2.0
